# api/docs/api-doc.yaml
swagger: "2.0"
info:
  description: 'Un api per un servizio di acquisto/vendita valute.'
  version: "1.0.0"
  title: "Exchange app"
basePath: "/api"
...
schemes:
  - "https"
  - "http"
paths:
  /users/signup:
    post:
      tags:
        - "users"
      summary: "Un api per registrare un utente"
      description: "Un api per registrare un utente"
      operationId: "signupUser"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "informazioni dell'utente da registrare"
          required: true
          schema:
            $ref: "#/definitions/SignupViewModel"
      responses:
        201:
          description: "Utente registrato con successo"
        400:
          description: "Errore nel corpo della richiesta"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
...